add APIUtils, prepare apification

+++ src/main/kotlin/org/bloqly/machine/model/NodeId.kt
@@ -8,7 +8,7 @@ data class NodeId(
 
     val host: String,
 
-    val port: Long
+    val port: Int
 
 ) : Serializable {
